requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
streamlit 
langchain
langchain_community
python-dotenv
pypdf
langchain_google_genai 